requirements.txt
requests==0.10.0
requests-oauth==0.4.0
Django==1.4
libxml2-python==2.6.9
